use DI

+++ Controllers/GitHubPackageStatusController.cs
@@ -7,23 +7,29 @@ using Microsoft.Extensions.Logging;
 
 namespace DotnetStatus.Controllers
 {
-    public class PackageStatusController : Controller
+    public class GitHubPackageStatusController : Controller
     {
-        private readonly ILogger<PackageStatusController> _log;
-        private readonly NuGetStatusService _statusService;
+        private readonly ILogger<GitHubPackageStatusController> _log;
+        private readonly IPackageStatusService _statusService;
 
-        public PackageStatusController(ILogger<PackageStatusController> log)
+        public GitHubPackageStatusController(
+            IPackageStatusService packageStatusService,
+            ILogger<GitHubPackageStatusController> log)
         {
-            _statusService = new NuGetStatusService();
+            _statusService = packageStatusService;
             _log = log;
         }
-        
+
         // api/status/gh/{*path}
-        public async Task<IActionResult> Get()
+        public async Task<IActionResult> GetGithub()
         {
             var route = Request.Path.Value.Split('/').ToList();
             
-            // api/status/gh/{*path} -> {*path}
+            // api/status/gh/user/project/path..csproj
+            if(route.Count < 6)
+                return BadRequest();
+
+            // api/status/gh/{*path} -> {*path} => 
             route.RemoveRange(0, 4);
 
             var formattedRoute = string.Join("/", route);
@@ -33,8 +39,11 @@ namespace DotnetStatus.Controllers
             var link = GetGithubLink(formattedRoute);
 
             _log.LogInformation($"link: {link}");
-            
+
             var result = await _statusService.GetStatusAsync(link);
+
+            if (result == null)
+                return NotFound();
 
             return Ok(result);
         }
